limit access to other contacts

+++ delete.php
@@ -28,10 +28,19 @@ if (!$contact) {
    echo("HTTP 404 NOT FOUND");
    return;
  }
-$statement = $mysqli->prepare("DELETE FROM contacts WHERE id = (?) ");
-$statement->bind_param('i', $id);
-$statement->execute();
-header("Location: home.php");
+
+ if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+  http_response_code(403);
+  echo ("HTTP 403 UNAUTHORIZED");
+  return;
+ } else {
+  $statement = $mysqli->prepare("DELETE FROM contacts WHERE id = (?) ");
+  $statement->bind_param('i', $id);
+  $statement->execute();
+  header("Location: home.php");
+ }
+
+
